Move package for models, constants, and types up
