+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/Constants.scala
@@ -1,0 +1,5 @@
+package io.github.shogowada.scala.jsonrpc
+
+object Constants {
+  val JsonRpc = "2.0"
+}
