--- shared/src/main/scala/io/github/shogowada/scala/jsonrpc/api/JsonRpcMethod.scala
@@ -1,5 +1,0 @@
-package io.github.shogowada.scala.jsonrpc.api
-
-import scala.annotation.StaticAnnotation
-
-class JsonRpcMethod(val name: String = "") extends StaticAnnotation
