+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/client/JsonRpcSingleMethodClient.scala
@@ -1,7 +1,7 @@
 package io.github.shogowada.scala.jsonrpc.client
 
-import io.github.shogowada.scala.jsonrpc.models.Types.Id
-import io.github.shogowada.scala.jsonrpc.models._
+import io.github.shogowada.scala.jsonrpc.Models.{JsonRpcErrorResponse, JsonRpcNotification, JsonRpcRequest, JsonRpcResponse}
+import io.github.shogowada.scala.jsonrpc.Types.Id
 import io.github.shogowada.scala.jsonrpc.serializers.JsonSerializer
 
 import scala.concurrent.{Future, Promise}
@@ -18,7 +18,7 @@ class JsonRpcSingleMethodClient[SERIALIZER[_], DESERIALIZER[_], PARAMS, ERROR, R
   private type ErrorOrResult = Either[JsonRpcErrorResponse[ERROR], JsonRpcResponse[RESULT]]
 
   def send(request: JsonRpcRequest[PARAMS]): Future[ErrorOrResult] = {
-//    jsonSerializer.serialize(request)
+    //    jsonSerializer.serialize(request)
     None
         .map(json => send(request.id, json))
         .getOrElse(Future.failed(new IllegalArgumentException(s"$request could not be serialized into JSON.")))
@@ -31,7 +31,7 @@ class JsonRpcSingleMethodClient[SERIALIZER[_], DESERIALIZER[_], PARAMS, ERROR, R
   }
 
   def send(notification: JsonRpcNotification[PARAMS]): Unit = {
-//    jsonSerializer.serialize(notification)
+    //    jsonSerializer.serialize(notification)
     None
         .foreach(jsonSender.send(_))
   }
@@ -45,16 +45,16 @@ class JsonRpcSingleMethodClient[SERIALIZER[_], DESERIALIZER[_], PARAMS, ERROR, R
   }
 
   private def maybeGetResultAsRight(json: String): Option[ErrorOrResult] = {
-//    jsonSerializer.deserialize[JsonRpcResponse[RESULT]](json)
-//        .filter(response => response.jsonrpc == Constants.JsonRpc)
-//        .map(response => Right(response))
+    //    jsonSerializer.deserialize[JsonRpcResponse[RESULT]](json)
+    //        .filter(response => response.jsonrpc == Constants.JsonRpc)
+    //        .map(response => Right(response))
     None
   }
 
   private def maybeGetErrorAsLeft(json: String): Option[ErrorOrResult] = {
-//    jsonSerializer.deserialize[JsonRpcErrorResponse[ERROR]](json)
-//        .filter(response => response.jsonrpc == Constants.JsonRpc)
-//        .map(response => Left(response))
+    //    jsonSerializer.deserialize[JsonRpcErrorResponse[ERROR]](json)
+    //        .filter(response => response.jsonrpc == Constants.JsonRpc)
+    //        .map(response => Left(response))
     None
   }
