--- shared/src/main/scala/io/github/shogowada/scala/jsonrpc/server/JsonReceiver.scala
@@ -1,15 +1,0 @@
-package io.github.shogowada.scala.jsonrpc.server
-
-import io.github.shogowada.scala.jsonrpc.serializers.JsonSerializer
-
-import scala.concurrent.Future
-import scala.language.higherKinds
-
-trait JsonReceiver {
-  def receive[SERIALIZER[_], DESERIALIZER[_]]
-  (
-      json: String
-  )(
-      jsonSerializer: JsonSerializer[SERIALIZER, DESERIALIZER]
-  ): Future[Option[String]]
-}
