+++ shared/src/main/scala/io/github/shogowada/scala/jsonrpc/server/JsonRpcSingleMethodServer.scala
@@ -1,7 +1,8 @@
 package io.github.shogowada.scala.jsonrpc.server
 
-import io.github.shogowada.scala.jsonrpc.models.Types.{JsonRpcNotificationMethod, JsonRpcRequestMethod}
-import io.github.shogowada.scala.jsonrpc.models._
+import io.github.shogowada.scala.jsonrpc.Models.JsonRpcRequest
+import io.github.shogowada.scala.jsonrpc.Types
+import io.github.shogowada.scala.jsonrpc.Types.{JsonRpcNotificationMethod, JsonRpcRequestMethod}
 import io.github.shogowada.scala.jsonrpc.serializers.JsonSerializer
 
 import scala.concurrent.Future
@@ -11,21 +12,21 @@ class JsonRpcRequestSingleMethodServer[PARAMS, ERROR, RESULT]
 (
     methodName: String,
     method: JsonRpcRequestMethod[PARAMS, ERROR, RESULT]
-) extends JsonReceiver {
+) {
 
   import scala.concurrent.ExecutionContext.Implicits.global
 
-  override def receive[SERIALIZER[_], DESERIALIZER[_]]
+  def receive[SERIALIZER[_], DESERIALIZER[_]]
   (
       json: String
   )(
       jsonSerializer: JsonSerializer[SERIALIZER, DESERIALIZER]
   ): Future[Option[String]] = {
-//    jsonSerializer.deserialize[JsonRpcRequest[PARAMS]](json)
-//        .filter(request => request.jsonrpc == Constants.JsonRpc)
-//        .filter(request => request.method == methodName)
-//        .map(request => handle(request)(jsonSerializer))
-//        .getOrElse(Future(None))
+    //    jsonSerializer.deserialize[JsonRpcRequest[PARAMS]](json)
+    //        .filter(request => request.jsonrpc == Constants.JsonRpc)
+    //        .filter(request => request.method == methodName)
+    //        .map(request => handle(request)(jsonSerializer))
+    //        .getOrElse(Future(None))
     Future(None)
   }
 
@@ -35,10 +36,10 @@ class JsonRpcRequestSingleMethodServer[PARAMS, ERROR, RESULT]
   )(
       jsonSerializer: JsonSerializer[SERIALIZER, DESERIALIZER]
   ): Future[Option[String]] = {
-//    method(request).map {
-//      case Right(result: JsonRpcResponse[RESULT]) => jsonSerializer.serialize(result)
-//      case Left(error: JsonRpcErrorResponse[ERROR]) => jsonSerializer.serialize(error)
-//    }
+    //    method(request).map {
+    //      case Right(result: JsonRpcResponse[RESULT]) => jsonSerializer.serialize(result)
+    //      case Left(error: JsonRpcErrorResponse[ERROR]) => jsonSerializer.serialize(error)
+    //    }
     Future(None)
   }
 }
@@ -47,24 +48,24 @@ class JsonRpcNotificationSingleMethodServer[PARAMS]
 (
     methodName: String,
     method: JsonRpcNotificationMethod[PARAMS]
-) extends JsonReceiver {
+) {
 
   import scala.concurrent.ExecutionContext.Implicits.global
 
-  override def receive[SERIALIZER[_], DESERIALIZER[_]]
+  def receive[SERIALIZER[_], DESERIALIZER[_]]
   (
       json: String
   )(
       jsonSerializer: JsonSerializer[SERIALIZER, DESERIALIZER]
   ): Future[Option[String]] = {
     Future(
-//      jsonSerializer.deserialize[JsonRpcNotification[PARAMS]](json)
-//          .filter(notification => notification.jsonrpc == Constants.JsonRpc)
-//          .filter(notification => notification.method == methodName)
-//          .flatMap(notification => {
-//            method(notification)
-//            None
-//          })
+      //      jsonSerializer.deserialize[JsonRpcNotification[PARAMS]](json)
+      //          .filter(notification => notification.jsonrpc == Constants.JsonRpc)
+      //          .filter(notification => notification.method == methodName)
+      //          .flatMap(notification => {
+      //            method(notification)
+      //            None
+      //          })
       None
     )
   }
